refactor, update shop owner & transfer info
