+++ src/components/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './order/table'
+export * from './order/calculate-total'
